Presets
Added Presets, which are basically starting states of the map. No Preset is the previous default. Four Mountains will place mountain tiles in the center of each quadrant. Central Lake will place a large block of deep water in the center.

Automatic Speed can now be lower than 1.

Binary Trees now set value to the inserted object if value is null, which should really only be the case for the first insertion in No Preset. Should probably be optimized at some point in the future.