Proportions
Added proportion weights to tile types, that influence the proportional distribution of each tile type. While influencing, do not expect even huge maps to stick to the proportions, as tile type links and their weights heavily influence how tiles are placed. As an example, a single mountain must be surrounded by 4 forests in the current default.

Fixed a bug introduced in the previous commit that broke FullPropagation.