Added Desert
Added desert which can only border sand with a very small weight from sand. With the same proportions as mountains by default, this creates a pattern of large swathes of desert once generated (typically), since the proportional factor will be large once a desert has been generated.

With this change, I had to change how repairs are made. Now calculates the graph-distance between each neighbor of the tile that needs to be repaired. Theoretically, this could be influenced by link weight, but to keep it simple I have opted to not do that. This repair mechanism should be entropic for most (if not all) continuous tile type setups.